// resources/js/interfaces/generated/dump-value-types.ts
/*
 * This file is auto-generated.
 * Don't update it manually, otherwise, your changes will be lost.
 * To update the file run `php bin/intellisense`.
 *
 * Generated at: 2026-01-02T00:36:13+00:00.
 */

export enum DumpValueType {
    Object = 'object',
    Array = 'array',
    String = 'string',
    Constant = 'constant',
    Uninitialized = 'uninitialized',
    Number = 'number',
    Closure = 'closure',
    Unknown = 'unknown',
}